Remove code insights logic and import module
