--- pipe/codeinsights/README.md
@@ -1,3 +1,0 @@
-# Code Insights
-This Python 3 package is responsible for interfacing with Bitbuckets code insights feature.
-It is currently internal to the `aligent/owasp-dependency-check-pipe` but will likely be extracted at some point in the future.
