--- pipe/codeinsights/__init__.py
@@ -1,1 +1,0 @@
-from .codeinsights import CodeInsights
